LRAC-5406 bug: companyId can be 0 after unmapping, disable won't work correctly

+++ modules/apps/analytics/analytics-settings-impl/src/main/java/com/liferay/analytics/settings/internal/configuration/AnalyticsConfigurationTrackerImpl.java
@@ -85,13 +85,9 @@ public class AnalyticsConfigurationTrackerImpl
 
 	@Override
 	public void deleted(String pid) {
-		_unmapPid(pid);
-
 		long companyId = getCompanyId(pid);
 
-		if (companyId == CompanyConstants.SYSTEM) {
-			return;
-		}
+		_unmapPid(pid);
 
 		_disable(companyId);
 	}
@@ -337,10 +333,14 @@ public class AnalyticsConfigurationTrackerImpl
 
 	private void _disable(long companyId) {
 		try {
-			_analyticsMessageLocalService.deleteAnalyticsMessages(companyId);
+			if (companyId != CompanyConstants.SYSTEM) {
+				_analyticsMessageLocalService.deleteAnalyticsMessages(
+					companyId);
 
-			_deleteAnalyticsAdmin(companyId);
-			_deleteSAPEntry(companyId);
+				_deleteAnalyticsAdmin(companyId);
+				_deleteSAPEntry(companyId);
+			}
+
 			_disableAuthVerifier();
 		}
 		catch (Exception exception) {
